add single dot browse view controller

+++ HeyYou/Dot.m
@@ -36,10 +36,10 @@
     dot.color = dotDict[@"color"];
     dot.stars = dotDict[@"stars"];
     dot.username = dotDict[@"username_id"];
-    long latitude = [dotDict[@"latitude"] doubleValue];
-    long longitude = [dotDict[@"longitude"] doubleValue];
-    NSLog(@"Latitude: %ld", latitude);
-    NSLog(@"Longitude: %ld", longitude);
+    double latitude = [dotDict[@"latitude"] doubleValue];
+    double longitude = [dotDict[@"longitude"] doubleValue];
+    NSLog(@"Latitude: %f", latitude);
+    NSLog(@"Longitude: %f", longitude);
     dot.location = CLLocationCoordinate2DMake(latitude, longitude);
     [tempArray addObject:dot];
   }
